Separate arguments for MemSQL and handled interruptions in binlog replication.
Summary:
Now you can specify a separate host, username, password, etc.
for MySQL and MemSQL. There's not much of a reason to flush the binlog,
so that option was removed. Now, when replicating each query, it stores the current binlog position on memsql. If it's interrupted while executing the query, the next time you restart ditto, it will read from the query it was interrupted at.

Test Plan: Tested replication with files in memsql/memsqltest

Reviewers: ankur, nikita, adam, pieguy, alex, vasili

Differential Revision: http://grizzly.memsql.com/D2812